backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
httpx==0.27.2
google-generativeai==0.8.3
python-dotenv==1.0.1
pydantic==2.9.2
semgrep>=1.66.0
bandit>=1.7.7
PyJWT>=2.8.0
cryptography>=41.0.0
